fix(strava): handle large upload IDs (qint64) + log upload response
Strava upload IDs exceed 32-bit int, so parseIdJsonStravaUploadObject's
toInt() truncated them to 0 — the status poll then hit /uploads/0 (404) and
both upload paths reported 'Failed' even though the upload was accepted.
Read the id as 64-bit and thread qint64 through checkUploadStatus / the
stravaUploadID fields. Also log the upload response and guard against a
missing id (2xx error/duplicate body).

+++ src/app/util.h
@@ -101,7 +101,7 @@ public:
     static QString parseJsonObjectVersion(const QString &data);
     static bool isVersionNewer(const QString &currentVersion, const QString &latestVersion);
     static void parseJsonStravaObject(QString data);
-    static int parseIdJsonStravaUploadObject(QString data);
+    static qint64 parseIdJsonStravaUploadObject(QString data);
     static int parseStravaUploadStatus(QString data);
 
     static QList<Sensor> parseJsonSensorList(QString data);
